refactor: extract remaining migrations
